requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
supabase==2.0.0
python-dotenv==1.0.0
marshmallow==3.20.0
werkzeug==3.0.1
bcrypt==4.0.1
gunicorn==21.2.0
pytest==7.4.0
pytest-flask==1.2.0
email-validator
